+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 - fix: default user-agent is now received from installed webview
 - updated: translations
 - new: permanent night mode (optional) #825
-- new: help button in Tab-Dialog
+- new: help button in Tab-Dialog (WIKI entry is in work)
+- new: help button in Filter-Dialog (WIKI entry is in work)
 
 ### v 8.8